--- Assets/LowPolygon/PolygonAncientEmpire/URP_ExtractMe.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b62d4366145c29048bfdfaca5f62afe6
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
