--- Assets/LowPolygon/PolygonFarm/URP_ExtractMe.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9287c40817ed80a42ab121b841a03381
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
