+++ Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
@@ -48,6 +48,7 @@ MonoBehaviour:
   - RpcMovePlayer
   - asd
   - dsds
+  - SetOwner
   DisableAutoOpenWizard: 1
   ShowSettings: 1
   DevRegionSetOnce: 1
