+++ Assets/Scripts/PlayerInformation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fd29f95ddc3dfeb40a05a09ffbf9db0a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
